Split test.cc
