+++ test/test-main.cc
@@ -1,0 +1,2 @@
+﻿#define CATCH_CONFIG_MAIN
+#include "catch.hh"
